fix: 进度条精度修复 + 云盘进度条全宽显示
1. SyncTaskState: overallProgress/progress 改用 Double 精度计算，
   消除大文件上传时百分比 0.1%~0.2% 的跳回抖动
2. FileEntryRow: 云盘模式改为垂直排列（文件信息+进度条各占一行），
   进度条 90% + 百分比 10%，填满卡片宽度
3. SyncStatusBar: 进度条 weight(0.9f) + 文字 weight(0.1f) 替代
   原来的 weight(1f) + Spacer 方案

+++ APP/Models/others/src/main/java/com/whmdg/mczj/tools/fileop/sync/SyncTaskState.kt
@@ -26,7 +26,7 @@ data class SyncFileProgress(
     val status: UploadStatus
 ) {
     val progress: Float
-        get() = if (totalBytes > 0) uploadedBytes.toFloat() / totalBytes else 0f
+        get() = if (totalBytes > 0) (uploadedBytes.toDouble() / totalBytes).toFloat() else 0f
 }
 
 /** 整体同步任务状态（运行时，不持久化） */
@@ -43,5 +43,5 @@ data class SyncTaskState(
     val fileProgress: Map<String, SyncFileProgress> = emptyMap()
 ) {
     val overallProgress: Float
-        get() = if (totalBytes > 0) transferredBytes.toFloat() / totalBytes else 0f
+        get() = if (totalBytes > 0) (transferredBytes.toDouble() / totalBytes).toFloat() else 0f
 }
